change rename, remove API
There is nothing special about rename_file and the DmaFile: as we operate
on a path alone it would work for all kinds of file, not just DmaFile.
It sat under DmaFile so that we could have some modularity but now that
we have scipio::io, the io module itself does that job.

remove has a similar issue: it can sit at the top level and we can leave
a version of it inside the DmaFile object that operates on the path that
we have.
